refactor(core): collapse duplicate commit validators
