+++ crates/loonfs-core/src/path/mod.rs
@@ -1,4 +1,8 @@
+//! Path parsing, reading, and mutation planning over namespace metadata.
+
 pub(crate) mod helpers;
 pub(crate) mod query;
 pub(crate) mod read;
 pub(crate) mod write;
+
+pub use helpers::validate_path_for_mutation;
